Add more tests for primitives and arrays (#6)
* Failed experiment


* Comment out broken tests


* Return empty slice


* Add error handling from MoonBit


* Add memblock info for passing tests


* Comment out broken cases


* Fix one test


* Comment other tests


* Update wasm


* e2e tests fail


* Update runtime testdata


* Add memblock info to tests


* Add more memblocks


* Add more array tests


* Add tests


* Uncomment passing tests


* Uncomment more working tests


* Uncomment passing tests


* Uncomment structs


* Uncomment working test


* More maps debugging


* Uncomment more working tests


* Add more array tests


* Improve myWasmMock


* Add more mock memory tests


* Update testdata.wasm


* Add byte array tests


* Add Char array tests


* Add Int16 array tests


* Add UInt16 array tests


* Start Array[Int] runtime tests


* Work on Array[Int]


* Fix Array[Int] tests


* Add Int extremes


* Add UInt extremes


* Add Int64 and UInt64 to array_tests.go


* Add Int64? and UInt64? runtime tests


* Update runtime


* Rearrange test files


* Add two runtime arrays tests


* Uncomment working test


* Fix Char tests


* Add Int16 tests


* Update runtime testdata


---------

+++ runtime/languages/moonbit/tests/tuples_test.go
@@ -21,7 +21,7 @@ func TestTupleOutput(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	expected := []any{123, true, "hello"}
+	expected := []any{int32(123), true, "hello"}
 
 	if result == nil {
 		t.Error("expected a result")
